+++ changelog.md
@@ -1,3 +1,10 @@
+# 0.3.0
+
+- Fix bug with images.
+- Add multiple modes and mode parameter.
+- Add toc parameter to automatically insert toc.
+- Fix: upload attachments before text update (this caused images to disappear after manually editing).
+
 # 0.2.0
 
 - Allow to input login and/or password during build